ITCH main done - everything working
